requirements.txt
boto3==1.7.60
boto==2.49.0
botocore==1.10.60
elasticsearch==2.1.0
enum34==1.1.2
jsonschema==2.5.1
logstash-handler==0.1.0
logstash_formatter==0.5.14
lxml==3.5.0
messytables
mock==1.3.0
nose-parameterized==0.5.0
nose==1.3.7
numpy==1.11.2
pandas==0.18.1
py4j==0.9
PyFunctional==0.7.1
python-dateutil==2.5.3
pytz==2015.7
redis==2.10.5
requests
scikit-learn
scipy
six==1.10.0
statsd==3.2.1
statsmodels==0.6.1
unittest2==1.1.0
urllib3